Move all options related code into a new module: pydoctor.options
