+++ pydoctor/test/test_commandline.py
@@ -5,7 +5,8 @@ import sys
 
 from pytest import raises
 
-from pydoctor import driver, model
+from pydoctor.options import Options
+from pydoctor import driver
 
 from . import CapSys
 
@@ -61,7 +62,7 @@ def test_projectbasedir_absolute(tmp_path: Path) -> None:
     Using L{Path.samefile()} is reliable, but requires an existing path.
     """
     assert tmp_path.is_absolute()
-    options = model.Options.from_namespace(driver.parse_args(["--project-base-dir", str(tmp_path)]))
+    options = Options.from_args(["--project-base-dir", str(tmp_path)])
     assert options.projectbasedirectory is not None
     assert options.projectbasedirectory.samefile(tmp_path)
     assert options.projectbasedirectory.is_absolute()
@@ -78,7 +79,7 @@ def test_projectbasedir_symlink(tmp_path: Path) -> None:
     link.symlink_to('target', target_is_directory=True)
     assert link.samefile(target)
 
-    options = model.Options.from_namespace(driver.parse_args(["--project-base-dir", str(link)]))
+    options = Options.from_args(["--project-base-dir", str(link)])
     assert options.projectbasedirectory is not None
     assert options.projectbasedirectory.samefile(target)
     assert options.projectbasedirectory.is_absolute()
@@ -91,7 +92,7 @@ def test_projectbasedir_relative() -> None:
     the options object.
     """
     relative = "projbasedirvalue"
-    options = model.Options.from_namespace(driver.parse_args(["--project-base-dir", relative]))
+    options = Options.from_args(["--project-base-dir", relative])
     assert options.projectbasedirectory is not None
     assert options.projectbasedirectory.is_absolute()
     assert options.projectbasedirectory.name == relative
@@ -102,8 +103,7 @@ def test_cache_enabled_by_default() -> None:
     """
     Intersphinx object caching is enabled by default.
     """
-    parser = driver.get_parser()
-    options = parser.parse_args([])
+    options = Options.defaults()
     assert options.enable_intersphinx_cache
 
 
@@ -112,10 +112,10 @@ def test_cli_warnings_on_error() -> None:
     The --warnings-as-errors option is disabled by default.
     This is the test for the long form of the CLI option.
     """
-    options = driver.parse_args([])
+    options = Options.defaults()
     assert options.warnings_as_errors == False
 
-    options = driver.parse_args(['--warnings-as-errors'])
+    options = Options.from_args(['--warnings-as-errors'])
     assert options.warnings_as_errors == True
 
 
@@ -123,7 +123,7 @@ def test_project_version_default() -> None:
     """
     When no --project-version is provided, it will default empty string.
     """
-    options = driver.parse_args([])
+    options = Options.defaults()
     assert options.projectversion == ''
 
 
@@ -131,7 +131,7 @@ def test_project_version_string() -> None:
     """
     --project-version can be passed as a simple string.
     """
-    options = driver.parse_args(['--project-version', '1.2.3.rc1'])
+    options = Options.from_args(['--project-version', '1.2.3.rc1'])
     assert options.projectversion == '1.2.3.rc1'
